backend/requirements.txt
fastapi
uvicorn[standard]
pandas
scikit-learn
joblib
python-dotenv
transformers
torch
huggingface-hub
pydantic
